ai-service/requirements.txt
fastapi
uvicorn
python-jose
passlib
python-multipart
openai
redis
pydantic
python-dotenv
astroid